+++ server/repository/settings-repository.go
@@ -61,6 +61,9 @@ var (
 	SettingFeatureAuthProviders           SettingName = SettingName{Name: "feature_auth_providers", Type: SettingTypeBool}
 	SettingFeatureRecurringBookings       SettingName = SettingName{Name: "feature_recurring_bookings", Type: SettingTypeBool}
 	SettingEnforceTOTP                    SettingName = SettingName{Name: "enforce_totp", Type: SettingTypeBool}
+	SettingKioskSecret                    SettingName = SettingName{Name: "kiosk_access_secret", Type: SettingTypeString}
+	SettingKioskModeEnabled               SettingName = SettingName{Name: "kiosk_mode_enabled", Type: SettingTypeBool}
+	SettingFeatureKioskMode               SettingName = SettingName{Name: "feature_kiosk_mode", Type: SettingTypeBool}
 )
 
 var settingsRepository *SettingsRepository
@@ -259,6 +262,8 @@ func (r *SettingsRepository) InitDefaultSettingsForOrg(organizationID string) er
 		"($1, '"+SettingFeatureNoUserLimit.Name+"', '0'), "+
 		"($1, '"+SettingFeatureCustomDomains.Name+"', '0'), "+
 		"($1, '"+SettingFeatureGroups.Name+"', '0'), "+
+		"($1, '"+SettingFeatureKioskMode.Name+"', '0'), "+
+		"($1, '"+SettingKioskModeEnabled.Name+"', '0'), "+
 		"($1, '"+SettingAllowAnyUser.Name+"', '1'), "+
 		"($1, '"+SettingDailyBasisBooking.Name+"', '0'), "+
 		"($1, '"+SettingNoAdminRestrictions.Name+"', '0'), "+
@@ -284,7 +289,8 @@ func (r *SettingsRepository) InitDefaultSettingsForOrg(organizationID string) er
 		"($1, '"+SettingBookingRetentionEnabled.Name+"', '0'), "+
 		"($1, '"+SettingBookingRetentionDays.Name+"', '365'), "+
 		"($1, '"+SettingSubjectDefault.Name+"', '"+strconv.Itoa(SettingSubjectDefaultOptional)+"'), "+
-		"($1, '"+SettingEnforceTOTP.Name+"', '0') "+
+		"($1, '"+SettingEnforceTOTP.Name+"', '0'), "+
+		"($1, '"+SettingKioskModeEnabled.Name+"', '0') "+
 		"ON CONFLICT (organization_id, name) DO NOTHING",
 		organizationID)
 	return err