+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.85.1](https://github.com/seatsurfing/seatsurfing/compare/v1.85.0...v1.85.1) (2026-04-19)
+
+
+### Bug Fixes
+
+* do not show existing client secret when editing auth provider ([#2021](https://github.com/seatsurfing/seatsurfing/issues/2021)) ([0136788](https://github.com/seatsurfing/seatsurfing/commit/0136788a9370caf6a353151b222bcfa9c7919eb9))
+* fix unique name check when updating existing auth provider ([#2020](https://github.com/seatsurfing/seatsurfing/issues/2020)) ([77084e5](https://github.com/seatsurfing/seatsurfing/commit/77084e5eb0e3efa7a0d46baaa0247941422eccb2))
+
+## [1.85.0](https://github.com/seatsurfing/seatsurfing/compare/v1.84.1...v1.85.0) (2026-04-19)
+
+
+### Features
+
+* kiosk mode ([#2012](https://github.com/seatsurfing/seatsurfing/issues/2012)) ([0ba8cbf](https://github.com/seatsurfing/seatsurfing/commit/0ba8cbf9a89f8a259a7d68b2dce19ebed1177d60))
+
+
+### Bug Fixes
+
+* make sure auth provider name is unique per organization ([#2016](https://github.com/seatsurfing/seatsurfing/issues/2016)) ([d9c7baf](https://github.com/seatsurfing/seatsurfing/commit/d9c7bafa15dd762a6cadb5bd1e27c48a4e718abd))
+* move template buttons for auth providers to top ([#2018](https://github.com/seatsurfing/seatsurfing/issues/2018)) ([6d17df3](https://github.com/seatsurfing/seatsurfing/commit/6d17df3c8cafc38ced484f4c442734c62f9230e4))
+
 ## [1.84.1](https://github.com/seatsurfing/seatsurfing/compare/v1.84.0...v1.84.1) (2026-04-17)